Refact: split large component file
